+++ libs/common/src/lib/upload/create-link/create-link.component.ts
@@ -57,7 +57,9 @@ export class CreateLinkComponent {
   hasValidKey = this.standaloneService.hasValidKey;
   pendingResourcesLimit = PENDING_RESOURCES_LIMIT;
   extractConfigEnabled = this.features.unstable.extractConfig;
+  splitConfigEnabled = this.features.unstable.splitConfig;
   extractStrategy?: string;
+  splitStrategy?: string;
   updateOptionsExpander = 0;
 
   get invalid() {
@@ -99,6 +101,7 @@ export class CreateLinkComponent {
                 formValue.css_selector,
                 formValue.xpath,
                 this.extractStrategy,
+                this.splitStrategy,
                 this.langCode.value,
               ),
             ),
@@ -113,6 +116,7 @@ export class CreateLinkComponent {
               formValue.css_selector,
               formValue.xpath,
               this.extractStrategy,
+              this.splitStrategy,
               this.langCode.value,
             ),
           ]);
@@ -132,6 +136,7 @@ export class CreateLinkComponent {
                       row.css_selector,
                       row.xpath,
                       this.extractStrategy,
+                      this.splitStrategy,
                       this.langCode.value,
                     ),
                   ),
@@ -179,6 +184,7 @@ export class CreateLinkComponent {
     css_selector?: string | null,
     xpath?: string | null,
     extract_strategy?: string,
+    split_strategy?: string,
     language?: string,
   ): Observable<{ uuid: string }> {
     return isCloudFile
@@ -192,6 +198,7 @@ export class CreateLinkComponent {
           undefined,
           undefined,
           extract_strategy,
+          split_strategy,
           language,
         );
   }