+++ libs/common/src/lib/upload/upload-sitemap/upload-sitemap.component.ts
@@ -61,7 +61,9 @@ export class UploadSitemapComponent {
   localstorage: { key: string; value: string }[] = [];
   standalone = this.standaloneService.standalone;
   extractConfigEnabled = this.features.unstable.extractConfig;
+  splitConfigEnabled = this.features.unstable.splitConfig;
   extractStrategy?: string;
+  splitStrategy?: string;
 
   constructor(
     public modal: ModalRef,
@@ -97,8 +99,9 @@ export class UploadSitemapComponent {
                     this.cleanParameters(this.cookies),
                     this.cleanParameters(this.localstorage),
                     this.extractStrategy,
+                    this.splitStrategy,
                   )
-                : this.uploadService.createCloudFileResource(link, this.selectedLabels, this.extractStrategy),
+                : this.uploadService.createCloudFileResource(link, this.selectedLabels, this.extractStrategy, this.splitStrategy),
             ),
           ),
         ),